fix: Removed use of mapfile, fixed multi-disk
Removed the use of the bash `mapfile` command which is not compatible
with some older versions of bash.  This should be more forgiving on the
install disks that can be used for installation.

Also fixed a bug where a multi-disk NON-encrypted setup was erroring
because the second disk was still being encrypted.

Resolves: #56