fix: Fixed automatic disk selection
The old method had a number of edge case issues and one major issue (it
would sometimes select a volume that was already mounted).  The new
method removes any drives that are already mounted as candidates.  This
will take care of the boot medium as well as loop devices and things
like that.

Resolves: #64