fix: Permanent fix to missing debootstrap script
The Ubuntu team seems to have stopped updating the debootstrap package
to include scripts (symlinks really) for the latest releases such as
jammy, lunar, and beyond.  So, I will now create those symbolic links
manually following the pattern of pointing them all to the 'gutsy'
script.  For debian, which is usually not a problem, I would instead
point at the 'debian-common' script.