fix: Fixed script selecting Ventoy disks
We need to filter out hotplug (USB) drives from the list of "acceptable"
disks to install to.  This does mean you will not be able to use this
script to install an OS to a bootable USB volume.  However, this is
much safer as I was continually wiping partitions on the booted USB
volumes.