feat: Adding bootstrapper config
Attempting a new method for initializing environment variables for a deb-install run. Instead of a
bunch of separate scripts that are all essentially copies of each other, this will be the
"universal" script that handles 80% of the key values. It allows farming out to a separate script
which only needs to export variables and have no other logic for more sophisticated configurations.