fix: Protecting all lsb_release calls from error msg
Given that LSB has actually been discontinued, some versions of Debian and Ubuntu will produce an
error message "No LSB modules are available". It gets sent to standard error, and so we want
to redirect that to avoid the message from interrupting the rest of the text processing we might
be doing.